requirements.txt
Django==4.1.3
django_debug_toolbar==3.8.1
python-decouple==3.6
django-filter==22.1